Fix board 1 multiball glitch
The separators and rollovers for the top lanes have been shifted to the left by 1.5 dp. This guarantees that the ball doesn't get stuck while exiting the onramp when multiball is activated. The upper drop targets have also been realigned to remain parallel to the newly positioned box above the drop targets. This change close #26.